requirements.txt
fastapi
uvicorn[standard]
torch
transformers
Pillow
opencv-python
python-multipart